:construction: WIP: display notification items and toggle read/all-read state with http requests
This commit is the first time that mdb-react-ui-kit [1] is used, since it makes the handling of Modals much easier (opening, closing, monitoring that state in the Context). Greate Success.

[1]: https://mdbootstrap.com/docs/react/components/modal

Also, browser hot-reload is implemented when using `npm run watch`. I'm not quite sure if that will stay, but we can always remove it later.

Sorry for large commit message and big commit, here's a potato: 🥔